Client 관련 Dao, Dto 통합
