+++ cope/src/cope/beans/client/ClientDto.java
@@ -1,19 +1,21 @@
 package cope.beans.client;
 
-import java.sql.Date;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+import cope.beans.utils.DateUtils;
 
 public class ClientDto {
 	private int clientNo;
-	private String clientId;
-	private String clientPw;
-	private String clientNick;
-	private String clientEmail;
+	private String clientId, clientPw, clientNick, clientEmail;
 	private short clientBirthYear;
 	private String clientGrade;
-	private Date clientUnlockDate;
+	private Date clientUnlockDate; // 정지 해제 날짜
+	
 	public ClientDto() {
 		super();
 	}
+
 	public int getClientNo() {
 		return clientNo;
 	}
@@ -23,18 +25,21 @@ public class ClientDto {
 	public String getClientId() {
 		return clientId;
 	}
+	// 정규 표현식 필요
 	public void setClientId(String clientId) {
 		this.clientId = clientId;
 	}
 	public String getClientPw() {
 		return clientPw;
 	}
+	// 정규 표현식 필요
 	public void setClientPw(String clientPw) {
 		this.clientPw = clientPw;
 	}
 	public String getClientNick() {
 		return clientNick;
 	}
+	// 정규 표현식 필요
 	public void setClientNick(String clientNick) {
 		this.clientNick = clientNick;
 	}
@@ -44,26 +49,40 @@ public class ClientDto {
 	public void setClientEmail(String clientEmail) {
 		this.clientEmail = clientEmail;
 	}
-	
-	public String getClientGrade() {
-		return clientGrade;
-	}
-	public void setClientGrade(String clientGrade) {
-		this.clientGrade = clientGrade;
-	}
-	public Date getClientUnlockDate() {
-		return clientUnlockDate;
-	}
-	public void setClientUnlockDate(Date clientUnlockDate) {
-		this.clientUnlockDate = clientUnlockDate;
-	}
 	public short getClientBirthYear() {
 		return clientBirthYear;
 	}
 	public void setClientBirthYear(short clientBirthYear) {
 		this.clientBirthYear = clientBirthYear;
 	}
-	
-	
-	
+	public String getClientGrade() {
+		return clientGrade;
+	}
+	public String getClientGradeKorean() {
+		if (clientGrade.equals("normal")) return "일반";
+		else if (clientGrade.equals("super")) return "관리자";
+		else return "미상";
+	}
+	// 조건식 필요
+	public void setClientGrade(String clientGrade) {
+		this.clientGrade = clientGrade;
+	}
+	public Date getClientUnlockDate() {
+		return clientUnlockDate;
+	}
+	public String getClientUnlockDateString() {
+		SimpleDateFormat simpleDateformat = new SimpleDateFormat("yyyy년 MM월 dd일 HH:mm:ss");				
+		return simpleDateformat.format(clientUnlockDate);
+	}
+	public void setClientUnlockDate(Date clientUnlockDate) {
+		this.clientUnlockDate = clientUnlockDate;
+	}
+	// 정지해제 날짜 설정 및 갱신
+	public void setClientUnlockDateRefresh(Date clientUnlockDate) throws Exception {
+		this.clientUnlockDate = clientUnlockDate;
+		DateUtils dateUtils = new DateUtils();
+		if (clientUnlockDate != null && dateUtils.compareWithSysdate(clientUnlockDate)) {
+			this.clientUnlockDate = null;
+		}
+	}
 }
