--- cope/src/cope/beans/client/ClientsoDto.java
@@ -1,44 +1,0 @@
-package cope.beans.client;
-
-public class ClientsoDto {
-	private int clientNo;
-	private String clientNick;
-	private String clientId;
-	private String clientPw;
-	private short clientBirth;
-	public ClientsoDto() {
-		super();
-	}
-	public int getClientNo() {
-		return clientNo;
-	}
-	public void setClientNo(int clientNo) {
-		this.clientNo = clientNo;
-	}
-	public String getClientNick() {
-		return clientNick;
-	}
-	public void setClientNick(String clientNick) {
-		this.clientNick = clientNick;
-	}
-	public String getClientId() {
-		return clientId;
-	}
-	public void setClientId(String clientId) {
-		this.clientId = clientId;
-	}
-	public String getClientPw() {
-		return clientPw;
-	}
-	public void setClientPw(String clientPw) {
-		this.clientPw = clientPw;
-	}
-	public short getClientBirth() {
-		return clientBirth;
-	}
-	public void setClientBirth(short clientBirth) {
-		this.clientBirth = clientBirth;
-	}
-
-}
-//주석
